Skip serialization of none values for the http api action payload
